async stream options: use shared pointer for pased retry policy
