+++ test/common/router/retry_state_impl_test.cc
@@ -56,8 +56,7 @@ public:
   }
 
   void setup(Http::RequestHeaderMap& request_headers) {
-
-    state_ = RetryStateImpl::create(policy_, request_headers, cluster_, &virtual_cluster_,
+    state_ = RetryStateImpl::create(*policy_, request_headers, cluster_, &virtual_cluster_,
                                     route_stats_context_, factory_context_, dispatcher_,
                                     Upstream::ResourcePriority::Default);
   }
@@ -166,7 +165,7 @@ public:
   void TearDown() override { cleanupOutstandingResources(); }
 
   Event::SimulatedTimeSystem test_time_;
-  NiceMock<TestRetryPolicy> policy_;
+  std::shared_ptr<NiceMock<TestRetryPolicy>> policy_ = TestRetryPolicy::createMock();
   NiceMock<Upstream::MockClusterInfo> cluster_;
   TestVirtualCluster virtual_cluster_;
   Stats::IsolatedStoreImpl stats_store_;
@@ -502,7 +501,7 @@ TEST_F(RouterRetryStateImplTest, PolicyRetriable4xxReset) {
 }
 
 TEST_F(RouterRetryStateImplTest, RetriableStatusCodes) {
-  policy_.retriable_status_codes_.push_back(409);
+  policy_->retriable_status_codes_.push_back(409);
   verifyPolicyWithRemoteResponse("retriable-status-codes", "409", false /* is_grpc */);
 }
 
@@ -533,7 +532,7 @@ TEST_F(RouterRetryStateImplTest, NoRetryUponTooEarlyStatusCodeWithDownstreamEarl
 }
 
 TEST_F(RouterRetryStateImplTest, RetriableStatusCodesUpstreamReset) {
-  policy_.retriable_status_codes_.push_back(409);
+  policy_->retriable_status_codes_.push_back(409);
   Http::TestRequestHeaderMapImpl request_headers{{"x-envoy-retry-on", "retriable-status-codes"}};
   setup(request_headers);
   EXPECT_TRUE(state_->enabled());
@@ -606,13 +605,13 @@ TEST_F(RouterRetryStateImplTest, RetriableStatusCodesHeader) {
 // Test that when 'retriable-headers' policy is set via request header, certain configured headers
 // trigger retries.
 TEST_F(RouterRetryStateImplTest, RetriableHeadersPolicySetViaRequestHeader) {
-  policy_.retry_on_ = RetryPolicy::RETRY_ON_5XX;
+  policy_->retry_on_ = RetryPolicy::RETRY_ON_5XX;
 
   Protobuf::RepeatedPtrField<envoy::config::route::v3::HeaderMatcher> matchers;
   auto* matcher = matchers.Add();
   matcher->set_name("X-Upstream-Pushback");
 
-  policy_.retriable_headers_ =
+  policy_->retriable_headers_ =
       Http::HeaderUtility::buildHeaderMatcherVector(matchers, factory_context_);
 
   // No retries based on response headers: retry mode isn't enabled.
@@ -644,7 +643,7 @@ TEST_F(RouterRetryStateImplTest, RetriableHeadersPolicySetViaRequestHeader) {
 // Test that when 'retriable-headers' policy is set via retry policy configuration,
 // configured header matcher conditions trigger retries.
 TEST_F(RouterRetryStateImplTest, RetriableHeadersPolicyViaRetryPolicyConfiguration) {
-  policy_.retry_on_ = RetryPolicy::RETRY_ON_RETRIABLE_HEADERS;
+  policy_->retry_on_ = RetryPolicy::RETRY_ON_RETRIABLE_HEADERS;
 
   Protobuf::RepeatedPtrField<envoy::config::route::v3::HeaderMatcher> matchers;
 
@@ -664,7 +663,7 @@ TEST_F(RouterRetryStateImplTest, RetriableHeadersPolicyViaRetryPolicyConfigurati
   matcher4->mutable_range_match()->set_start(500);
   matcher4->mutable_range_match()->set_end(505);
 
-  policy_.retriable_headers_ =
+  policy_->retriable_headers_ =
       Http::HeaderUtility::buildHeaderMatcherVector(matchers, factory_context_);
 
   auto should_retry_with_response = [this](const Http::TestResponseHeaderMapImpl& response_headers,
@@ -784,7 +783,7 @@ TEST_F(RouterRetryStateImplTest, RetriableHeadersSetViaRequestHeader) {
 
 // Test merging retriable headers set via request headers and via config file.
 TEST_F(RouterRetryStateImplTest, RetriableHeadersMergedConfigAndRequestHeaders) {
-  policy_.retry_on_ = RetryPolicy::RETRY_ON_RETRIABLE_HEADERS;
+  policy_->retry_on_ = RetryPolicy::RETRY_ON_RETRIABLE_HEADERS;
 
   Protobuf::RepeatedPtrField<envoy::config::route::v3::HeaderMatcher> matchers;
 
@@ -794,7 +793,7 @@ TEST_F(RouterRetryStateImplTest, RetriableHeadersMergedConfigAndRequestHeaders) 
   matcher->mutable_string_match()->set_exact("200");
   matcher->set_invert_match(true);
 
-  policy_.retriable_headers_ =
+  policy_->retriable_headers_ =
       Http::HeaderUtility::buildHeaderMatcherVector(matchers, factory_context_);
 
   // No retries according to config.
@@ -856,7 +855,7 @@ TEST_F(RouterRetryStateImplTest, PolicyLimitedByRequestHeaders) {
   matcher2->set_name(":method");
   matcher2->mutable_string_match()->set_exact("HEAD");
 
-  policy_.retriable_request_headers_ =
+  policy_->retriable_request_headers_ =
       Http::HeaderUtility::buildHeaderMatcherVector(matchers, factory_context_);
 
   {
@@ -917,8 +916,8 @@ TEST_F(RouterRetryStateImplTest, PolicyLimitedByRequestHeaders) {
 }
 
 TEST_F(RouterRetryStateImplTest, RouteConfigNoRetriesAllowed) {
-  policy_.num_retries_ = 0;
-  policy_.retry_on_ = RetryPolicy::RETRY_ON_CONNECT_FAILURE;
+  policy_->num_retries_ = 0;
+  policy_->retry_on_ = RetryPolicy::RETRY_ON_CONNECT_FAILURE;
   setup();
 
   EXPECT_TRUE(state_->enabled());
@@ -935,8 +934,8 @@ TEST_F(RouterRetryStateImplTest, RouteConfigNoRetriesAllowed) {
 }
 
 TEST_F(RouterRetryStateImplTest, RouteConfigNoHeaderConfig) {
-  policy_.num_retries_ = 1;
-  policy_.retry_on_ = RetryPolicy::RETRY_ON_CONNECT_FAILURE;
+  policy_->num_retries_ = 1;
+  policy_->retry_on_ = RetryPolicy::RETRY_ON_CONNECT_FAILURE;
   Http::TestRequestHeaderMapImpl request_headers;
   setup(request_headers);
   EXPECT_TRUE(state_->enabled());
@@ -966,7 +965,7 @@ TEST_F(RouterRetryStateImplTest, NoAvailableRetries) {
 
 TEST_F(RouterRetryStateImplTest, MaxRetriesHeader) {
   // The max retries header will take precedence over the policy
-  policy_.num_retries_ = 4;
+  policy_->num_retries_ = 4;
   Http::TestRequestHeaderMapImpl request_headers{{"x-envoy-retry-on", "connect-failure"},
                                                  {"x-envoy-retry-grpc-on", "cancelled"},
                                                  {"x-envoy-max-retries", "3"}};
@@ -1011,8 +1010,8 @@ TEST_F(RouterRetryStateImplTest, MaxRetriesHeader) {
 }
 
 TEST_F(RouterRetryStateImplTest, Backoff) {
-  policy_.num_retries_ = 5;
-  policy_.retry_on_ = RetryPolicy::RETRY_ON_CONNECT_FAILURE;
+  policy_->num_retries_ = 5;
+  policy_->retry_on_ = RetryPolicy::RETRY_ON_CONNECT_FAILURE;
   Http::TestRequestHeaderMapImpl request_headers;
   setup(request_headers);
   EXPECT_TRUE(state_->enabled());
@@ -1082,10 +1081,10 @@ TEST_F(RouterRetryStateImplTest, Backoff) {
 
 // Test customized retry back-off intervals.
 TEST_F(RouterRetryStateImplTest, CustomBackOffInterval) {
-  policy_.num_retries_ = 10;
-  policy_.retry_on_ = RetryPolicy::RETRY_ON_CONNECT_FAILURE;
-  policy_.base_interval_ = std::chrono::milliseconds(100);
-  policy_.max_interval_ = std::chrono::milliseconds(1200);
+  policy_->num_retries_ = 10;
+  policy_->retry_on_ = RetryPolicy::RETRY_ON_CONNECT_FAILURE;
+  policy_->base_interval_ = std::chrono::milliseconds(100);
+  policy_->max_interval_ = std::chrono::milliseconds(1200);
   Http::TestRequestHeaderMapImpl request_headers;
   setup(request_headers);
   EXPECT_TRUE(state_->enabled());
@@ -1134,9 +1133,9 @@ TEST_F(RouterRetryStateImplTest, CustomBackOffInterval) {
 
 // Test the default maximum retry back-off interval.
 TEST_F(RouterRetryStateImplTest, CustomBackOffIntervalDefaultMax) {
-  policy_.num_retries_ = 10;
-  policy_.retry_on_ = RetryPolicy::RETRY_ON_CONNECT_FAILURE;
-  policy_.base_interval_ = std::chrono::milliseconds(100);
+  policy_->num_retries_ = 10;
+  policy_->retry_on_ = RetryPolicy::RETRY_ON_CONNECT_FAILURE;
+  policy_->base_interval_ = std::chrono::milliseconds(100);
   Http::TestRequestHeaderMapImpl request_headers;
   setup(request_headers);
   EXPECT_TRUE(state_->enabled());
@@ -1197,7 +1196,7 @@ TEST_F(RouterRetryStateImplTest, ParseRateLimitedResetInterval) {
   reset_header_2->set_name("X-RateLimit-Reset");
   reset_header_2->set_format(envoy::config::route::v3::RetryPolicy::UNIX_TIMESTAMP);
 
-  policy_.reset_headers_ = ResetHeaderParserImpl::buildResetHeaderParserVector(reset_headers);
+  policy_->reset_headers_ = ResetHeaderParserImpl::buildResetHeaderParserVector(reset_headers);
 
   // Failure case: Matches reset header (seconds) but exceeds max_interval (>5min)
   {
@@ -1263,8 +1262,8 @@ TEST_F(RouterRetryStateImplTest, RateLimitedRetryBackoffStrategy) {
   reset_header->set_name("Retry-After");
   reset_header->set_format(envoy::config::route::v3::RetryPolicy::SECONDS);
 
-  policy_.num_retries_ = 4;
-  policy_.reset_headers_ = ResetHeaderParserImpl::buildResetHeaderParserVector(reset_headers);
+  policy_->num_retries_ = 4;
+  policy_->reset_headers_ = ResetHeaderParserImpl::buildResetHeaderParserVector(reset_headers);
 
   Http::TestRequestHeaderMapImpl request_headers{{"x-envoy-retry-on", "5xx"}};
   setup(request_headers);
@@ -1320,8 +1319,8 @@ TEST_F(RouterRetryStateImplTest, RateLimitedRetryBackoffStrategy) {
 }
 
 TEST_F(RouterRetryStateImplTest, HostSelectionAttempts) {
-  policy_.host_selection_max_attempts_ = 2;
-  policy_.retry_on_ = RetryPolicy::RETRY_ON_CONNECT_FAILURE;
+  policy_->host_selection_max_attempts_ = 2;
+  policy_->retry_on_ = RetryPolicy::RETRY_ON_CONNECT_FAILURE;
 
   setup();
