+++ test/common/router/router_test.cc
@@ -1033,7 +1033,7 @@ public:
 // Also verify retry options predicates work.
 TEST_F(RouterTest, EnvoyAttemptCountInRequestUpdatedInRetries) {
   auto retry_options_predicate = std::make_shared<MockRetryOptionsPredicate>();
-  callbacks_.route_->route_entry_.retry_policy_.retry_options_predicates_.emplace_back(
+  callbacks_.route_->route_entry_.retry_policy_->retry_options_predicates_.emplace_back(
       retry_options_predicate);
 
   setIncludeAttemptCountInRequest(true);
@@ -1912,7 +1912,7 @@ TEST_F(RouterTest, UpstreamTimeoutWithAltResponse) {
 TEST_F(RouterTest, UpstreamPerTryIdleTimeout) {
   InSequence s;
 
-  callbacks_.route_->route_entry_.retry_policy_.per_try_idle_timeout_ =
+  callbacks_.route_->route_entry_.retry_policy_->per_try_idle_timeout_ =
       std::chrono::milliseconds(3000);
 
   // This pattern helps ensure that we're actually invoking the callback.
@@ -1983,7 +1983,7 @@ TEST_F(RouterTest, UpstreamPerTryIdleTimeout) {
 TEST_F(RouterTest, UpstreamPerTryIdleTimeoutSuccess) {
   InSequence s;
 
-  callbacks_.route_->route_entry_.retry_policy_.per_try_idle_timeout_ =
+  callbacks_.route_->route_entry_.retry_policy_->per_try_idle_timeout_ =
       std::chrono::milliseconds(3000);
 
   NiceMock<Http::MockRequestEncoder> encoder;
@@ -2188,7 +2188,7 @@ TEST_F(RouterTest, UpstreamPerTryTimeoutExcludesNewStream) {
 // canceled). Also verify retry options predicates work.
 TEST_F(RouterTest, HedgedPerTryTimeoutFirstRequestSucceeds) {
   auto retry_options_predicate = std::make_shared<MockRetryOptionsPredicate>();
-  callbacks_.route_->route_entry_.retry_policy_.retry_options_predicates_.emplace_back(
+  callbacks_.route_->route_entry_.retry_policy_->retry_options_predicates_.emplace_back(
       retry_options_predicate);
 
   enableHedgeOnPerTryTimeout();
@@ -2679,7 +2679,7 @@ TEST_F(RouterTest, BadHeadersDroppedIfPreviousRetryScheduled) {
 // has sent any of the body. Also verify retry options predicates work.
 TEST_F(RouterTest, RetryRequestBeforeBody) {
   auto retry_options_predicate = std::make_shared<MockRetryOptionsPredicate>();
-  callbacks_.route_->route_entry_.retry_policy_.retry_options_predicates_.emplace_back(
+  callbacks_.route_->route_entry_.retry_policy_->retry_options_predicates_.emplace_back(
       retry_options_predicate);
 
   NiceMock<Http::MockRequestEncoder> encoder1;
@@ -5861,7 +5861,7 @@ TEST(RouterFilterUtilityTest, FinalTimeout) {
   }
   {
     NiceMock<MockRouteEntry> route;
-    route.retry_policy_.per_try_timeout_ = std::chrono::milliseconds(7);
+    route.retry_policy_->per_try_timeout_ = std::chrono::milliseconds(7);
     EXPECT_CALL(route, timeout()).WillOnce(Return(std::chrono::milliseconds(10)));
     Http::TestRequestHeaderMapImpl headers{{"x-envoy-upstream-rq-timeout-ms", "15"}};
     TimeoutData timeout = FilterUtility::finalTimeout(route, headers, true, false, false, false);
@@ -5874,7 +5874,7 @@ TEST(RouterFilterUtilityTest, FinalTimeout) {
   }
   {
     NiceMock<MockRouteEntry> route;
-    route.retry_policy_.per_try_timeout_ = std::chrono::milliseconds(10);
+    route.retry_policy_->per_try_timeout_ = std::chrono::milliseconds(10);
     EXPECT_CALL(route, timeout()).WillOnce(Return(std::chrono::milliseconds(0)));
     Http::TestRequestHeaderMapImpl headers;
     TimeoutData timeout = FilterUtility::finalTimeout(route, headers, true, false, false, false);
@@ -5887,7 +5887,7 @@ TEST(RouterFilterUtilityTest, FinalTimeout) {
   }
   {
     NiceMock<MockRouteEntry> route;
-    route.retry_policy_.per_try_timeout_ = std::chrono::milliseconds(7);
+    route.retry_policy_->per_try_timeout_ = std::chrono::milliseconds(7);
     EXPECT_CALL(route, timeout()).WillOnce(Return(std::chrono::milliseconds(10)));
     Http::TestRequestHeaderMapImpl headers{{"x-envoy-upstream-rq-timeout-ms", "15"},
                                            {"x-envoy-upstream-rq-per-try-timeout-ms", "5"}};
@@ -6040,7 +6040,7 @@ TEST(RouterFilterUtilityTest, FinalTimeout) {
     NiceMock<MockRouteEntry> route;
     EXPECT_CALL(route, maxGrpcTimeout())
         .WillRepeatedly(Return(absl::optional<std::chrono::milliseconds>(0)));
-    route.retry_policy_.per_try_timeout_ = std::chrono::milliseconds(7);
+    route.retry_policy_->per_try_timeout_ = std::chrono::milliseconds(7);
     Http::TestRequestHeaderMapImpl headers{{"content-type", "application/grpc"},
                                            {"grpc-timeout", "1000m"},
                                            {"x-envoy-upstream-rq-timeout-ms", "15"}};
@@ -6056,7 +6056,7 @@ TEST(RouterFilterUtilityTest, FinalTimeout) {
     NiceMock<MockRouteEntry> route;
     EXPECT_CALL(route, maxGrpcTimeout())
         .WillRepeatedly(Return(absl::optional<std::chrono::milliseconds>(0)));
-    route.retry_policy_.per_try_timeout_ = std::chrono::milliseconds(7);
+    route.retry_policy_->per_try_timeout_ = std::chrono::milliseconds(7);
     Http::TestRequestHeaderMapImpl headers{{"content-type", "application/grpc"},
                                            {"grpc-timeout", "1000m"},
                                            {"x-envoy-upstream-rq-timeout-ms", "15"},
@@ -6701,7 +6701,7 @@ TEST_F(RouterTest, Http3DisabledForHttp11Proxies) {
 
 TEST_F(RouterTest, ExpectedUpstreamTimeoutUpdatedDuringRetries) {
   auto retry_options_predicate = std::make_shared<MockRetryOptionsPredicate>();
-  callbacks_.route_->route_entry_.retry_policy_.retry_options_predicates_.emplace_back(
+  callbacks_.route_->route_entry_.retry_policy_->retry_options_predicates_.emplace_back(
       retry_options_predicate);
 
   setIncludeAttemptCountInRequest(true);
